+++ src/app/(main)/content/_components/ContentItem.tsx
@@ -1,0 +1,5 @@
+const ContentItem = () => {
+  return <article className="w-full p-4 flex flex-col gap-[10px]"></article>;
+};
+
+export default ContentItem;
