+++ src/app/(main)/content/_components/ContentList.tsx
@@ -1,0 +1,5 @@
+const ContentList = () => {
+  return <div className="w-full flex flex-col gap-4"></div>;
+};
+
+export default ContentList;
