+++ src/app/(main)/content/_components/MainContent.tsx
