+++ src/app/(main)/section/_components/MainSection.tsx
