+++ src/app/page.tsx
@@ -5,9 +5,9 @@ import Image from "next/image";
 import { Container } from "@/components";
 import { TopBar } from "@/components/common/topBar";
 import { useScrollBlur } from "@/utils/useScrollBlur";
-import MainSection from "./main/_components/section/MainSection";
 import BNB from "@/components/common/bnb";
-import MainContent from "./main/_components/content/MainContent";
+import MainContent from "./(main)/content/_components/MainContent";
+import MainSection from "./(main)/section/_components/MainSection";
 
 export default function Home() {
   const containerRef = useRef<HTMLElement>(null);
